+++ .gitignore
@@ -5,3 +5,4 @@
 .DS_Store
 /.vagrant
 yarn.lock
+snippets.js
