+++ src/config.json
@@ -1,5 +1,6 @@
 {
 	"port": 8081,
 	"bodyLimit": "100kb",
-	"corsHeaders": ["Link"]
+	"corsHeaders": ["Link"],
+	"appSpace": "/tmp/dd-agent"
 }
