+++ src/db.js
@@ -1,4 +1,4 @@
 export default callback => {
-	// connect to a database if needed, then pass it to `callback`:
-	callback();
-}
+  // connect to a database if needed, then pass it to `callback`:
+  callback();
+};
