+++ src/middleware/index.js
@@ -1,9 +1,9 @@
-import { Router } from 'express';
+// import { Router } from 'express';
 
-export default ({ config, db }) => {
-	let routes = Router();
+// export default ({ config, db }) => {
+//   const routes = Router();
 
-	// add middleware here
+//   // add middleware here
 
-	return routes;
-}
+//   return routes;
+// };
