--- src/models/facets.js
@@ -1,3 +1,0 @@
-// our example model is just an Array
-const facets = [];
-export default facets;
